requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
albucore
albumentations==1.4.14
altair==5.3.0
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
av==12.3.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
cachetools==5.4.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloths-segmentation==0.0.2
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cycler==0.12.0
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docker-pycreds==0.4.0
efficientnet_pytorch==0.7.1
et_xmlfile==2.0.0
eval_type_backport==0.2.0
executing==2.0.1
faiss-cpu==1.10.0
fastjsonschema==2.19.1
filelock==3.15.4
flash==1.0.3
Flask==3.0.0
flask-ngrok==0.0.25
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.6
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.35.1
ipykernel==6.28.0
ipython==8.19.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.2
joblib==1.4.2
jpeg4py==0.1.4
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.1
jupyter_server==2.12.1
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-huggingface==0.1.2
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
langsmith==0.3.0
lazy_loader==0.4
lightning-utilities==0.11.6
lpips==0.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.25.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.13.1
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
notebook==7.0.6
notebook_shim==0.2.3
numpy
openai==1.61.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.15
overrides==7.4.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
Pillow==10.0.1
platformdirs==4.1.0
pretrainedmodels==0.7.4
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==5.27.2
psutil==5.9.7
pure-eval==0.2.2
pyarrow==17.0.0
pycparser==2.21
pydantic==2.8.2
pydantic-settings==2.7.1
pydantic_core==2.20.1
pydeck==0.9.1
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytorch-lightning==2.4.0
pytorch-toolbelt==0.6.3
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.16.2
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.2
sentence-transformers==3.4.1
sentry-sdk==2.13.0
setproctitle==1.3.3
setuptools==73.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.37
stack-data==0.6.3
streamlit==1.36.0
sympy==1.13.2
tenacity==8.5.0
tensorboardX==2.6.2.2
terminado==0.18.0
threadpoolctl==3.5.0
tifffile==2024.8.10
tiktoken==0.8.0
timm==0.9.2
tinycss2==1.2.1
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
torchgeometry==0.1.2
torchmetrics==1.4.1
torchvision==0.19.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.0
transformers==4.48.3
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
wandb==0.17.7
watchdog==4.0.1
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.0
widgetsnbextension==4.0.9
yarl==1.9.4
zstandard==0.23.0